LPS-98481 Inline ScriptingExecutorTestCase to GroovyExecutorTest. The commit LPS-96677(80ab259b5bbb3024906b446858a7c5b6b59cc2ed) removed other scripting and only GroovyExecutorTest extends ScriptingExecutorTestCase.
